rnd: Change parameter type in randombytes_{weak,strong}
Change parameter type for the length to size_t since this is what is
passed to them and is expected by read_exact().

+++ rnd.c
@@ -10,7 +10,7 @@
 
 static int fdw = -1;
 
-static void randombytes_weak(unsigned char *x, unsigned long long xlen)
+static void randombytes_weak(unsigned char *x, size_t xlen)
 {
 	int ret;
 
@@ -40,7 +40,7 @@ static void randombytes_weak(unsigned char *x, unsigned long long xlen)
 	}
 }
 
-static void randombytes_strong(unsigned char *x, unsigned long long xlen)
+static void randombytes_strong(unsigned char *x, size_t xlen)
 {
 	int fds, ret;
